Test minor refactoring
